demande + backendUpdate

+++ project/Backend/src/main/java/ma/emsi/repositories/DemandeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DemandeRepository extends JpaRepository<Demande, Long> {
+public interface DemandeRepository extends JpaRepository<Demande, Integer> {
 
-    Demande findById(long id);
+    Demande findById(int id);
 }
